requirements.txt
discord.py
python-dotenv
validators
lxml
